chore(ee): move automations + guards to core, drop src/pocketpaw/ee/ (Phase 2)
Phase 2 of the open-core split — final subpackage moves.

- Deleted the empty placeholder ee/pocketpaw_ee/automations/ (zero
  importers; the real automations engine already lived in core).
- src/pocketpaw/ee/automations/ -> src/pocketpaw/automations/ — the
  rule-based automation engine, relocated off the confusing
  pocketpaw.ee.* path onto the canonical pocketpaw.automations.
- src/pocketpaw/ee/guards/ -> src/pocketpaw/guards/ — RBAC/ABAC policy
  package, fully self-contained, same relocation.
- Removed the now-empty src/pocketpaw/ee/ directory.
- automations router moved from _EE_ROUTERS to _V1_ROUTERS — it's core
  now (its one pocketpaw_ee.api dep is a lazy in-function import,
  pre-existing debt for Phase 3).

ee/pocketpaw_ee/ now holds only: cloud, agent, audit, calendar, fleet,
api.py, and the three split router packages (fabric, instinct,
paw_print).
