--- ee/pocketpaw_ee/automations/__init__.py
@@ -1,4 +1,0 @@
-# Automations — time/data triggers for Paw OS.
-# Created: 2026-03-28 — Placeholder for future implementation.
-# "When inventory drops below 10, alert me."
-# "Every Monday, generate the weekly report pocket."
