+++ ee/pocketpaw_ee/cloud/_core/deps.py
@@ -22,9 +22,9 @@ from typing import Any
 
 from fastapi import Depends, HTTPException
 
-from pocketpaw.ee.guards.audit import log_denial
-from pocketpaw.ee.guards.deps import check_workspace_action
-from pocketpaw.ee.guards.rbac import Forbidden as GuardForbidden
+from pocketpaw.guards.audit import log_denial
+from pocketpaw.guards.deps import check_workspace_action
+from pocketpaw.guards.rbac import Forbidden as GuardForbidden
 from pocketpaw_ee.cloud._core.errors import Forbidden
 from pocketpaw_ee.cloud.auth import current_active_user
 from pocketpaw_ee.cloud.models.user import User
@@ -160,7 +160,7 @@ def require_plan_feature(feature: str) -> Callable[..., Coroutine[Any, Any, None
             ],
         )
     """
-    from pocketpaw.ee.guards.abac import PLAN_FEATURES
+    from pocketpaw.guards.abac import PLAN_FEATURES
 
     # Compute the minimum plan that unlocks this feature, for the error message.
     needed_plan = "enterprise"
