+++ ee/pocketpaw_ee/fleet/router.py
@@ -40,8 +40,8 @@ from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel, Field
 from soul_protocol.engine.journal import Journal
 
-from pocketpaw.ee.guards.deps import check_workspace_action
-from pocketpaw.ee.guards.rbac import Forbidden as GuardForbidden
+from pocketpaw.guards.deps import check_workspace_action
+from pocketpaw.guards.rbac import Forbidden as GuardForbidden
 from pocketpaw_ee.cloud.auth import current_active_user
 from pocketpaw_ee.cloud.models.user import User
 from pocketpaw_ee.fleet import (
