+++ src/pocketpaw/api/v1/__init__.py
@@ -63,9 +63,11 @@ _V1_ROUTERS: list[tuple[str, str, str]] = [
     # ee.instinct.router; both forward semantically to this one.
     ("pocketpaw.audit.runtime_router", "router", "Runtime Audit"),
     # Moved to OSS core in the open-core split (Phase 2) — journal-backed
-    # retrieval + widget projections, no multi-tenant cloud dependency.
+    # retrieval + widget projections + rule-based automations, no
+    # multi-tenant cloud dependency.
     ("pocketpaw.retrieval.router", "router", "Retrieval"),
     ("pocketpaw.widget.router", "router", "Widgets"),
+    ("pocketpaw.automations.router", "router", "Automations"),
 ]
 
 # Enterprise API routes (require ee/ module) — skipped silently when ee/ is absent.
@@ -73,7 +75,6 @@ _EE_ROUTERS: list[tuple[str, str, str]] = [
     ("pocketpaw_ee.fabric.router", "router", "Fabric"),
     ("pocketpaw_ee.fleet.router", "router", "Fleet"),
     ("pocketpaw_ee.instinct.router", "router", "Instinct"),
-    ("pocketpaw.ee.automations.router", "router", "Automations"),
 ]
