--- src/pocketpaw/ee/__init__.py
@@ -1,11 +1,0 @@
-# PocketPaw Enterprise Edition — gated features for team/business/enterprise plans.
-# Created: 2026-03-30
-# Updated: 2026-04-10 — Added guards module (RBAC + ABAC).
-# License: FSL 1.1 (Functional Source License)
-#
-# Module map:
-#   guards/       — RBAC + ABAC authorization primitives and FastAPI dependencies
-#   automations/  — Trigger-action workflows (time, event, webhook)
-#   fabric/       — Workspace resource mesh and cross-pocket orchestration
-#   instinct/     — Proactive agent behaviors and autonomous action policies
-#   audit/        — Enterprise audit trail, compliance exports, and retention policies
